+++ activerecord/test/cases/adapters/mysql2/boolean_test.rb
@@ -46,8 +46,8 @@ class Mysql2BooleanTest < ActiveRecord::TestCase
     assert_equal 1, attributes["archived"]
     assert_equal "1", attributes["published"]
 
-    assert_equal "t", @connection.type_cast(true, boolean_column)
-    assert_equal "t", @connection.type_cast(true, string_column)
+    assert_equal 1, @connection.type_cast(true, boolean_column)
+    assert_equal 1, @connection.type_cast(true, string_column)
   end
 
   test "test type casting without emulated booleans" do
@@ -60,7 +60,7 @@ class Mysql2BooleanTest < ActiveRecord::TestCase
     assert_equal "1", attributes["published"]
 
     assert_equal 1, @connection.type_cast(true, boolean_column)
-    assert_equal "t", @connection.type_cast(true, string_column)
+    assert_equal 1, @connection.type_cast(true, string_column)
   end
 
   test "with booleans stored as 1 and 0" do